updated readme, moved types dir
